+++ biblioteca-4/tablas/db-personas/modificar-3.php
@@ -20,7 +20,7 @@ if (hayErrores("modificar-3")) {
 
 [$nombre, $apellidos, $dni, $id] = compruebaAvisosIndividuales("modificar-3", "personas", "nombre", "apellidos", "dni", "id");
 
-compruebaAvisosGenerales("modificar-3", "todosVacios", "nombre", "apellidos", "dni");
+compruebaAvisosGenerales("modificar-3", "todosVaciosMenosPrimero", "id", "nombre", "apellidos", "dni");
 
 compruebaAvisosGenerales("modificar-3", "yaExisteRegistroConOtroId", "personas", "nombre", "apellidos", "dni", "id");